Changed BLIND_FUNC to accept multiple arguments
Multiple arguments as a function happens when a specific templatization
of a function with multiple, comma separated, template parameters is
used. The preprocessor interprets the eg “bar<int, int>” as two
arguments: “bat<int” and “int>”. By using __VA_ARGS__, we can capture
all arguments, and “perfectly forward” the templetized function name to
the preprocessing macro.

Added a test case for the above scenario.